requirements.txt
tensorflow==2.16.1
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.16.1  # Match the TensorFlow version
matplotlib
numpy
